Fix compilation errors and refactor filter components
- Fixed missing using directive in GoogleCalendarServiceTests
- Fixed Shouldly type casting error in PartnerFilteringPerformanceTests
- Added missing Output property to IntegrationContext base class
- Fixed MudChip type inference compilation errors in Home.razor
- Completely refactored PartnerFilterComponent with proper RenderFragment syntax
- Simplified complex render fragment structure to resolve Razor compilation issues
- Updated OptimizedFilterService to align with PartnerFilterCriteria model
- Maintained core filtering functionality while fixing structural issues

All compilation errors resolved - solution builds successfully

+++ src/EventServer.Tests/PartnerFilteringPerformanceTests.cs
@@ -314,7 +314,7 @@ public class PartnerFilteringPerformanceTests : IntegrationContext
         var failedRequests = results.Where(r => !r.Success).ToList();
         
         // Most requests should succeed
-        successfulRequests.Count.ShouldBeGreaterThan(results.Length * 0.8, 
+        successfulRequests.Count.ShouldBeGreaterThan((int)(results.Length * 0.8), 
             $"At least 80% of concurrent requests should succeed. Success: {successfulRequests.Count}, Failed: {failedRequests.Count}");
         
         // Overall time should be reasonable (not much longer than the longest individual request)
